export and remove duplicate function
